fix header overlay, update open string border size, update file structure, cleanup
