+++ src/components/Header.jsx
@@ -4,17 +4,23 @@ import { Text, View, StyleSheet, Pressable } from "react-native";
 import { Store } from "../../Store";
 import { theme } from "../utils/theme";
 
+
 const Header = () => {
-  const { globalState, showMenu, setShowMenu, showOptions, setShowOptions, dimensions } = useContext(Store);
+  const { globalState, showMenu, setShowMenu, showOptions, setShowOptions } = useContext(Store);
+  const overlay = () => showOptions && setShowOptions(false);
+  const fullScreen = { height: "100%", width: "100%" };
 
   return (
     <View style={styles.container}>
-      <Pressable style={styles.menuButton} onPress={() => setShowMenu(!showMenu)}>
-        <Text style={styles.menu}>Menu</Text>
-      </Pressable>
-      <Text style={styles.heading}>{globalState?.scale.title}</Text>
-      <Pressable style={styles.settingsButtonContainer} onPress={() => !showMenu && setShowOptions(!showOptions)}>
-        <Text style={[styles.settingsButton, showMenu && styles.disableOptions]}>● ● ●</Text>
+      <Pressable style={fullScreen} onPress={overlay}>
+        <Pressable style={styles.menuButton} onPress={() => setShowMenu(!showMenu)}>
+          <Text style={styles.menu}>Menu</Text>
+        </Pressable>
+        <Text style={styles.heading}>{globalState?.scale.title}</Text>
+        <Pressable style={styles.settingsButtonContainer} onPress={() => !showMenu && setShowOptions(!showOptions)}>
+          <Text style={[styles.settingsButton, showMenu && styles.disableOptions]}>● ● ●</Text>
+        </Pressable>
+        {showOptions && <View style={styles.overlay}/>}
       </Pressable>
     </View>
   );
@@ -22,10 +28,12 @@ const Header = () => {
 
 export default Header;
 
+export const headerHeight = 38;
+
 const styles = StyleSheet.create({
   container: {
     backgroundColor: theme.colors.blue,
-    height: 38,
+    height: headerHeight,
     position: "absolute",
     top: 0,
     width: "100%",
@@ -55,6 +63,13 @@ const styles = StyleSheet.create({
     zIndex: 100,
     width: 100,
     height: 50,
+  },
+  overlay: {
+    backgroundColor: theme.colors.overlay,
+    zIndex: 2000,
+    position: "absolute",
+    width: '100%',
+    height: headerHeight + 20,
   },
   settingsButton: {
     color: theme.colors.white,
