--- src/components/Neck-old.jsx
@@ -1,808 +1,0 @@
-import React, { useContext } from 'react'
-import { Pressable, View, StyleSheet } from 'react-native';
-import Svg, { Path, Text } from "react-native-svg";
-import { Store } from "../../Store";
-
-const NeckSVG = () => {
-  const { globalState } = useContext(Store);
-  const degrees = globalState.scale.degrees;
-  const firstString = globalState.strings[0];
-  const secondString = globalState.strings[1];
-  const thirdString = globalState.strings[2];
-  const fourthString = globalState.strings[3];
-  const fifthString = globalState.strings[4];
-  const sixthString = globalState.strings[5];
-
-  const getOffset = (fret, stringOffset) => {
-    let offset = Math.floor(fret) + stringOffset;
-    offset -= globalState.key.key_offset;
-    offset = offset < 0 ? offset + 12 : offset > 11 ? offset - 12 : offset;
-    return offset
-  }
-
-  const isNote = (fret, stringOffset) => degrees.includes(getOffset(fret, stringOffset)) ? true : false
-
-  const getScaleDegree = (fret, stringOffset) => {
-    const offset = getOffset(fret, stringOffset);
-    switch (offset) {
-      case 0:
-        return '1';
-        break;
-      case 1:
-        return 'b2';
-        break;
-      case 2:
-        return '2';
-        break;
-      case 3:
-        return 'b3';
-        break;
-      case 3.1:
-        return '#2';
-        break;
-      case 4:
-        return '3';
-        break;
-      case 5:
-        return '4';
-        break;
-      case 6:
-        return 'b5';
-        break;
-      case 6.1:
-        return '#4';
-        break;
-      case 7:
-        return '5';
-        break;
-      case 8:
-        return 'b6';
-        break;
-      case 8.1:
-        return '#5';
-        break;
-      case 9:
-        return '6';
-        break;
-      case 10:
-        return 'b7';
-        break;
-      case 11:
-        return '7';
-        break;
-    }
-  }
-
-
-  return (
-    <Svg xmlns="http://www.w3.org/2000/svg" width="100%" height="100%" fill="none">
-
-      {/* strings */}
-      <Path stroke="#000" strokeWidth="4" d="M39 14.5h798M39 50.5h798M39 86.5h798M39 122.5h798M40 158.5h798M39 194.5h798" />
-
-      {/* nut */}
-      <Path stroke="#000" strokeWidth="6" d="M40 196.5v-184" />
-
-      {/* frets */}
-      <Path stroke="#000" strokeWidth="4" d="M90 196.5v-184M139 196.5v-184M188 196.5v-184M237 196.5v-184M286 196.5v-184M335 196.5v-184M384 196.5v-184M433 196.5v-184M482 196.5v-184M531 196.5v-184M580 196.5v-184M629 196.5v-184M678 196.5v-184M727 196.5v-184M776 196.5v-184M825 196.5v-184" />
-
-      {/* anchor frets */}
-      {!globalState.options.hideAnchorFrets && (
-        <>
-          <Path fill="#000" d="M164 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M164 232.5a3 3 0 100-6 3 3 0 000 6z" />
-          <Path fill="#000" d="M262 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M262 232.5a3 3 0 100-6 3 3 0 000 6z" />
-          <Path fill="#000" d="M360 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M360 232.5a3 3 0 100-6 3 3 0 000 6z" />
-          <Path fill="#000" d="M458 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M458 232.5a3 3 0 100-6 3 3 0 000 6z" />
-          <Path fill="#000" d="M601 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M601 232.5a3 3 0 100-6 3 3 0 000 6z" />
-          <Path fill="#000" d="M611 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M611 232.5a3 3 0 100-6 3 3 0 000 6z" />
-          <Path fill="#000" d="M754 233a3.5 3.5 0 100-7 3.5 3.5 0 000 7z" />
-          <Path stroke="#000" d="M754 232.5a3 3 0 100-6 3 3 0 000 6z" />
-        </>
-      )}
-
-      {/* first string */}
-      {isNote(0, firstString) && (
-        <Path stroke="#656565" strokeWidth="2" d="M14 26.518c6.914 0 12.518-5.604 12.518-12.518 0-6.914-5.604-12.519-12.518-12.519C7.086 1.481 1.481 7.086 1.481 14S7.086 26.518 14 26.518z" />
-      )}
-      {isNote(1, firstString) && (
-        <>
-          <Path fill="#000" d="M65 26c6.627 0 12-5.373 12-12S71.627 2 65 2 53 7.373 53 14s5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M65 25.556c6.382 0 11.555-5.174 11.555-11.556 0-6.382-5.173-11.556-11.555-11.556S53.444 7.618 53.444 14c0 6.382 5.174 11.556 11.556 11.556z" />
-          <Text
-            fill="white"
-            stroke="white"
-            fontSize="20"
-            fontWeight="bold"
-            x="58"
-            y="20"
-          >
-            {getScaleDegree(1, firstString)}
-          </Text>
-        </>
-      )}
-      {isNote(2, firstString) && (
-        <>
-          <Path fill="#000" d="M114 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M114 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-          <Text
-            fill="white"
-            stroke="white"
-            fontSize="20"
-            fontWeight="bold"
-            x="100"
-            y="20"
-          >
-            {getScaleDegree(2, firstString)}
-          </Text>
-        </>
-      )}
-      {isNote(3, firstString) && (
-        <>
-          <Path fill="#000" d="M164 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M164 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-          <Text
-            fill="white"
-            stroke="white"
-            fontSize="20"
-            fontWeight="bold"
-            x="157"
-            y="20"
-          >
-            {getScaleDegree(3, firstString)}
-          </Text>
-        </>
-      )}
-      {isNote(4, firstString) && (
-        <>
-          <Path fill="#000" d="M212 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M212 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(5, firstString) && (
-        <>
-          <Path fill="#000" d="M261 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M261 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(6, firstString) && (
-        <>
-          <Path fill="#000" d="M310 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M310 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(7, firstString) && (
-        <>
-          <Path fill="#000" d="M359 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M359 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(8, firstString) && (
-        <>
-          <Path fill="#000" d="M408 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M408 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(9, firstString) && (
-        <>
-          <Path fill="#000" d="M457 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M457 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(10, firstString) && (
-        <>
-          <Path fill="#000" d="M506 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M506 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(11, firstString) && (
-        <>
-          <Path fill="#000" d="M555 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M555 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(12, firstString) && (
-        <>
-          <Path fill="#000" d="M604 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M604 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(13, firstString) && (
-        <>
-          <Path fill="#000" d="M653 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M653 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(14, firstString) && (
-        <>
-          <Path fill="#000" d="M703 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M703 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(15, firstString) && (
-        <>
-          <Path fill="#000" d="M752 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M752 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(16, firstString) && (
-        <>
-          <Path fill="#000" d="M801 26c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M801 25.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-
-      {/* second string */}
-      {isNote(0, secondString) && (
-        <Path stroke="#656565" strokeWidth="2" d="M14 62.519c6.914 0 12.518-5.605 12.518-12.519S20.915 37.481 14 37.481C7.086 37.481 1.481 43.087 1.481 50S7.086 62.519 14 62.519z" />
-      )}
-      {isNote(1, secondString) && (
-        <>
-          <Path fill="#000" d="M65 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M65 61.556c6.382 0 11.555-5.174 11.555-11.556 0-6.382-5.173-11.556-11.555-11.556S53.444 43.618 53.444 50c0 6.382 5.174 11.556 11.556 11.556z" />
-          <Text
-            fill="white"
-            stroke="white"
-            fontSize="20"
-            fontWeight="bold"
-            x="60"
-            y="56"
-          >
-            {getScaleDegree(1, secondString)}
-          </Text>
-        </>
-      )}
-      {isNote(2, secondString) && (
-        <>
-          <Path fill="#000" d="M114 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M114 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-          <Text
-            fill="white"
-            stroke="white"
-            fontSize="20"
-            fontWeight="bold"
-            x="106"
-            y="56"
-          >
-            {getScaleDegree(2, secondString)}
-          </Text>
-        </>
-      )}
-      {isNote(3, secondString) && (
-        <>
-          <Path fill="#000" d="M164 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M164 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-          <Text
-            fill="white"
-            stroke="white"
-            fontSize="20"
-            fontWeight="bold"
-            x="157"
-            y="56"
-          >
-            {getScaleDegree(3, secondString)}
-          </Text>
-        </>
-      )}
-      {isNote(4, secondString) && (
-        <>
-          <Path fill="#000" d="M212 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M212 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(5, secondString) && (
-        <>
-          <Path fill="#000" d="M261 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M261 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(6, secondString) && (
-        <>
-          <Path fill="#000" d="M310 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M310 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(7, secondString) && (
-        <>
-          <Path fill="#000" d="M359 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M359 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(8, secondString) && (
-        <>
-          <Path fill="#000" d="M408 63c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M408 62.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(9, secondString) && (
-        <>
-          <Path fill="#000" d="M457 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M457 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(10, secondString) && (
-        <>
-          <Path fill="#000" d="M506 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M506 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(11, secondString) && (
-        <>
-          <Path fill="#000" d="M555 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M555 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(12, secondString) && (
-        <>
-          <Path fill="#000" d="M604 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M604 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(13, secondString) && (
-        <>
-          <Path fill="#000" d="M653 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M653 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(14, secondString) && (
-        <>
-          <Path fill="#000" d="M703 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M703 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(15, secondString) && (
-        <>
-          <Path fill="#000" d="M752 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M752 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(16, secondString) && (
-        <>
-          <Path fill="#000" d="M801 62c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M801 61.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-
-      {/* third string */}
-      {isNote(0, thirdString) && (
-        <Path stroke="#656565" strokeWidth="2" d="M14 98.519c6.914 0 12.518-5.605 12.518-12.519S20.915 73.481 14 73.481C7.086 73.481 1.481 79.087 1.481 86S7.086 98.519 14 98.519z" />
-      )}
-      {isNote(1, thirdString) && (
-        <>
-          <Path fill="#000" d="M65 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M65 97.556c6.382 0 11.555-5.174 11.555-11.556 0-6.382-5.173-11.556-11.555-11.556S53.444 79.618 53.444 86c0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(2, thirdString) && (
-        <>
-          <Path fill="#000" d="M114 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M114 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(3, thirdString) && (
-        <>
-          <Path fill="#000" d="M164 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M164 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(4, thirdString) && (
-        <>
-          <Path fill="#000" d="M212 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M212 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(5, thirdString) && (
-        <>
-          <Path fill="#000" d="M261 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M261 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S254.618 97.556 261 97.556z" />
-        </>
-      )}
-      {isNote(6, thirdString) && (
-        <>
-          <Path fill="#000" d="M310 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M310 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S303.618 97.556 310 97.556z" />
-        </>
-      )}
-      {isNote(7, thirdString) && (
-        <>
-          <Path fill="#000" d="M359 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M359 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S352.618 97.556 359 97.556z" />
-        </>
-      )}
-      {isNote(8, thirdString) && (
-        <>
-          <Path fill="#000" d="M408 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M408 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S401.618 97.556 408 97.556z" />
-        </>
-      )}
-      {isNote(9, thirdString) && (
-        <>
-          <Path fill="#000" d="M457 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M457 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S450.618 97.556 457 97.556z" />
-        </>
-      )}
-      {isNote(10, thirdString) && (
-        <>
-          <Path fill="#000" d="M506 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M506 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S499.618 97.556 506 97.556z" />
-        </>
-      )}
-      {isNote(11, thirdString) && (
-        <>
-          <Path fill="#000" d="M555 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M555 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S548.618 97.556 555 97.556z" />
-        </>
-      )}
-      {isNote(12, thirdString) && (
-        <>
-          <Path fill="#000" d="M604 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M604 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S597.618 97.556 604 97.556z" />
-        </>
-      )}
-      {isNote(13, thirdString) && (
-        <>
-          <Path fill="#000" d="M653 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M653 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S646.618 97.556 653 97.556z" />
-        </>
-      )}
-      {isNote(14, thirdString) && (
-        <>
-          <Path fill="#000" d="M703 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M703 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S696.618 97.556 703 97.556z" />
-        </>
-      )}
-      {isNote(15, thirdString) && (
-        <>
-          <Path fill="#000" d="M752 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M752 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S745.618 97.556 752 97.556z" />
-        </>
-      )}
-      {isNote(16, thirdString) && (
-        <>
-          <Path fill="#000" d="M801 98c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M801 97.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.555-11.556-11.555-6.382 0-11.556 5.173-11.556 11.555S794.618 97.556 801 97.556z" />
-        </>
-      )}
-
-      {/* fourth string */}
-      {isNote(0, fourthString) && (
-        <Path stroke="#656565" strokeWidth="2" d="M14 134.519c6.914 0 12.518-5.605 12.518-12.519S20.915 109.481 14 109.481c-6.914 0-12.519 5.605-12.519 12.519S7.086 134.519 14 134.519z" />
-      )}
-      {isNote(1, fourthString) && (
-        <>
-          <Path fill="#000" d="M65 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M65 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(2, fourthString) && (
-        <>
-          <Path fill="#000" d="M114 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M114 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(3, fourthString) && (
-        <>
-          <Path fill="#000" d="M164 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M164 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(4, fourthString) && (
-        <>
-          <Path fill="#000" d="M212 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M212 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(5, fourthString) && (
-        <>
-          <Path fill="#000" d="M261 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M261 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(6, fourthString) && (
-        <>
-          <Path fill="#000" d="M310 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M310 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(7, fourthString) && (
-        <>
-          <Path fill="#000" d="M359 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M359 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(8, fourthString) && (
-        <>
-          <Path fill="#000" d="M408 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M408 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(9, fourthString) && (
-        <>
-          <Path fill="#000" d="M457 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M457 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(10, fourthString) && (
-        <>
-          <Path fill="#000" d="M506 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M506 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(11, fourthString) && (
-        <>
-          <Path fill="#000" d="M555 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M555 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(12, fourthString) && (
-        <>
-          <Path fill="#000" d="M604 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M604 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(13, fourthString) && (
-        <>
-          <Path fill="#000" d="M653 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M653 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(14, fourthString) && (
-        <>
-          <Path fill="#000" d="M703 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M703 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(15, fourthString) && (
-        <>
-          <Path fill="#000" d="M752 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M752 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(16, fourthString) && (
-        <>
-          <Path fill="#000" d="M801 134c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M801 133.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-
-      {/* fifth string */}
-      {isNote(0, fifthString) && (
-        <Path stroke="#656565" strokeWidth="2" d="M14 170.519c6.914 0 12.518-5.605 12.518-12.519S20.915 145.481 14 145.481c-6.914 0-12.519 5.605-12.519 12.519S7.086 170.519 14 170.519z" />
-      )}
-      {isNote(1, fifthString) && (
-        <>
-          <Path fill="#000" d="M65 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M65 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(2, fifthString) && (
-        <>
-          <Path fill="#000" d="M114 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M114 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(3, fifthString) && (
-        <>
-          <Path fill="#000" d="M164 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M164 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(4, fifthString) && (
-        <>
-          <Path fill="#000" d="M212 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M212 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(5, fifthString) && (
-        <>
-          <Path fill="#000" d="M261 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M261 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(6, fifthString) && (
-        <>
-          <Path fill="#000" d="M310 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M310 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(7, fifthString) && (
-        <>
-          <Path fill="#000" d="M359 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M359 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(8, fifthString) && (
-        <>
-          <Path fill="#000" d="M408 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M408 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(9, fifthString) && (
-        <>
-          <Path fill="#000" d="M457 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M457 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(10, fifthString) && (
-        <>
-          <Path fill="#000" d="M506 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M506 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(11, fifthString) && (
-        <>
-          <Path fill="#000" d="M555 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M555 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(12, fifthString) && (
-        <>
-          <Path fill="#000" d="M604 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M604 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(13, fifthString) && (
-        <>
-          <Path fill="#000" d="M653 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M653 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(14, fifthString) && (
-        <>
-          <Path fill="#000" d="M703 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M703 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(15, fifthString) && (
-        <>
-          <Path fill="#000" d="M752 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M752 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-      {isNote(16, fifthString) && (
-        <>
-          <Path fill="#000" d="M801 170c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M801 169.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.555 5.174-11.555 11.556 0 6.382 5.173 11.556 11.555 11.556z" />
-        </>
-      )}
-
-      {/* sixth string */}
-      {isNote(0, sixthString) && (
-        <Path stroke="#656565" strokeWidth="2" d="M14 206.519c6.914 0 12.518-5.605 12.518-12.519S20.915 181.481 14 181.481c-6.914 0-12.519 5.605-12.519 12.519S7.086 206.519 14 206.519z" />
-      )}
-      {isNote(1, sixthString) && (
-        <>
-          <Path fill="#000" d="M65 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M65 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(2, sixthString) && (
-        <>
-          <Path fill="#000" d="M114 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M114 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(3, sixthString) && (
-        <>
-          <Path fill="#000" d="M163 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path fill="#000" stroke="#000" strokeWidth="4" d="M163 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(4, sixthString) && (
-        <>
-          <Path fill="#000" d="M212 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M212 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(5, sixthString) && (
-        <>
-          <Path fill="#000" d="M261 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M261 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(6, sixthString) && (
-        <>
-          <Path fill="#000" d="M310 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M310 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(7, sixthString) && (
-        <>
-          <Path fill="#000" d="M359 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M359 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(8, sixthString) && (
-        <>
-          <Path fill="#000" d="M408 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M408 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(9, sixthString) && (
-        <>
-          <Path fill="#000" d="M457 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M457 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(10, sixthString) && (
-        <>
-          <Path fill="#000" d="M506 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M506 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(11, sixthString) && (
-        <>
-          <Path fill="#000" d="M555 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M555 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(12, sixthString) && (
-        <>
-          <Path fill="#000" d="M604 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M604 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(13, sixthString) && (
-        <>
-          <Path fill="#000" d="M653 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M653 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(14, sixthString) && (
-        <>
-          <Path fill="#000" d="M703 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M703 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(15, sixthString) && (
-        <>
-          <Path fill="#000" d="M752 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M752 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-      {isNote(16, sixthString) && (
-        <>
-          <Path fill="#000" d="M801 206c6.627 0 12-5.373 12-12s-5.373-12-12-12-12 5.373-12 12 5.373 12 12 12z" />
-          <Path stroke="#000" strokeWidth="4" d="M801 205.556c6.382 0 11.556-5.174 11.556-11.556 0-6.382-5.174-11.556-11.556-11.556-6.382 0-11.556 5.174-11.556 11.556 0 6.382 5.174 11.556 11.556 11.556z" />
-        </>
-      )}
-    </Svg>
-  )
-}
-
-
-
-const Neck = () => {
-  return (
-    <View style={{ width: 838, height: 233 }}>
-      {/* <Pressable onPress={()=>console.log('press 1')} style={styles.pressable1} />
-      <Pressable onPress={()=>console.log('press 2')} style={styles.pressable2} /> */}
-      <NeckSVG />
-    </View>
-  )
-}
-
-export default Neck;
-
-const styles = StyleSheet.create({
-  pressable1: {
-    backgroundColor: '#FF0000aa',
-    width: "20%",
-    height: "100%",
-    position: 'absolute',
-    zIndex: 2
-  },
-  pressable2: {
-    backgroundColor: '#00FF00aa',
-    width: "27%",
-    height: "100%",
-    position: 'absolute',
-    zIndex: 2,
-    left: "20%"
-  }
-})
