+++ src/components/Splash/TitleSVG.jsx
